pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "eventlet>=0.40.4",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-login>=0.6.3",
    "flask-socketio>=5.6.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "geoalchemy2>=0.18.1",
    "geojson>=3.2.0",
    "gunicorn>=25.0.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "reportlab>=4.4.9",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.46",
    "werkzeug>=3.1.5",
    "wtforms>=3.2.1",
]
